requirements.txt
pydantic>=1.10
mysql-connector-python>=8.0
